added check for before deleting / creating following users

+++ bibsonomy-database/src/main/java/org/bibsonomy/database/DBLogic.java
@@ -1670,7 +1670,9 @@ public class DBLogic implements LogicInterface {
 			param.setRequestedUserName(requestedUser.getName());
 
 			if (UserRelation.FOLLOWER_OF.equals(relation)) {
-				this.userDBManager.deleteFollowerOfUser(param, session);
+				if (this.userDBManager.isFollowerOfUser(loggedInUser, requestedUser, session)) {
+					this.userDBManager.deleteFollowerOfUser(param, session);
+				}
 			}
 			if (UserRelation.FRIEND_OF.equals(relation)) {
 				this.userDBManager.deleteFriendOfUser(param, session);
@@ -1697,7 +1699,9 @@ public class DBLogic implements LogicInterface {
 			param.setRequestedUserName(requestedUser.getName());
 
 			if (UserRelation.FOLLOWER_OF.equals(relation)) {
-				this.userDBManager.addFollowerOfUser(param, session);
+				if (this.userDBManager.isFollowerOfUser(loggedInUser, requestedUser, session) == false) {
+					this.userDBManager.addFollowerOfUser(param, session);
+				}
 			}
 		} finally {
 			session.close();
